Passing all update items tests
- Enabled as well external http tests (using bin/server and port 8100 + balancer)
- Instead of logging to syslog, just echoing the error when happens. Need some revision
- Added volume for supervisor logs to var/logs/supervisor
