+++ bin/server
@@ -34,7 +34,7 @@ $kernel = new \Mmoreram\BaseBundle\Kernel\BaseKernel(
     $oneBundleAppConfig->getConfig(),
     $oneBundleAppConfig->getRoutes(),
     $environment,
-    false,
+    true,
     realpath(__DIR__ . '/../var')
 );
 
@@ -138,7 +138,7 @@ $loop->run();
 function exceptionToSyslog(\Exception $e)
 {
     openlog("search-server", LOG_PID, LOG_LOCAL0);
-    syslog(LOG_ALERT, "[{$e->getFile()}] [{$e->getCode()}] ::: [{$e->getMessage()}]");
+    echo("[{$e->getFile()}] [{$e->getCode()}] ::: [{$e->getMessage()}]");
     closelog();
 }
 
@@ -153,6 +153,6 @@ function toSyslog(
     string $content
 ) {
     openlog("search-server", LOG_PID, LOG_LOCAL0);
-    syslog($level, $content);
+    echo($content);
     closelog();
 }
