+++ docker-compose.yml
@@ -3,7 +3,7 @@ version: '3.3'
 services:
     elasticsearch:
         image: docker.elastic.co/elasticsearch/elasticsearch:6.2.4
-        container_name: elasticsearch
+        container_name: apisearch_elasticsearch
         environment:
             - cluster.name=apisearch
             - http.host=0.0.0.0
@@ -18,19 +18,19 @@ services:
 
     redis:
         image: redis:4-alpine
-        container_name: redis
+        container_name: apisearch_redis
 
     haproxy:
-        build: ./docker/haproxy
-        container_name: haproxy
+        build: ${PWD}/docker/haproxy
+        container_name: apisearch_haproxy
         ports:
             - 8100:8100
         links:
             - apisearch
 
     apisearch:
-        build: ./docker/search-server
-        container_name: apisearch
+        build: ${PWD}/docker/search-server
+        container_name: apisearch_server
         depends_on:
             - elasticsearch
             - redis
@@ -41,7 +41,8 @@ services:
             - 8203:8203
             - 8204:8204
         volumes:
-            - .:/var/www/apisearch
+            - ${PWD}:/var/www/apisearch
+            - ${PWD}/var/logs/supervisor:/var/log/supervisor
         links:
             - redis
