+++ docker/search-server/app_deploy.yml
@@ -1,12 +1,12 @@
 rs_queue:
     server:
         redis:
-            host: redis
+            host: apisearch_redis
 
 apisearch_server:
     middleware_domain_events_service: apisearch_server.middleware.ignore_events
     middleware_logs_service: apisearch_server.middleware.ignore_logs
     cluster:
         docker_elasticsearch_service:
-            host: elasticsearch
+            host: apisearch_elasticsearch
             port: 9200
